+++ packages/core/chain/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vultisig/core-chain
 
+## 1.6.1
+
+### Patch Changes
+
+- [#431](https://github.com/vultisig/vultisig-sdk/pull/431) [`1132ae5`](https://github.com/vultisig/vultisig-sdk/commit/1132ae51f8e4d5b8ca8a1855af9ea51031b574e9) Thanks [@gomesalexandre](https://github.com/gomesalexandre)! - fix cosmos denom resolver picking wrong segment for 3-part factory denoms
+
 ## 1.6.0
 
 ### Minor Changes